requirements.txt
python-dotenv
langchain
langchain-community
langchain-anthropic
langchain-chroma
langchain-ollama
beautifulsoup4
#ollama
#langchain-ollama